+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [Unreleased]
 ### Added
+- RequestsData can use .input extensions
+
 ### Fixed
 ### Changed
 - requests status code is optional, defaults to 200